Add: App.class
In this commit we add the main class where we can see some basic things about syntaxes and some basic things about OOP (create an object, access to the attributes of the object, setting the attributes that can be set of the object.  We have other class to create objects "Computer" and some basics things about this.

Resume:
-OOP
-Classes/Objects
-Class Attributes
-Class Methods
-Constructors
-Modifiers