Add: Inner Class
In this commit was added an inner class to Cat class (outter class), and we can see that a public inner class can access to attributes and methods of the outter class, but is necessary to create an object of the outter class to create an object of the inner class.

In the static inner class just like static attributes and methods, a static inner class does not have access to members of the outer class.

Resume:
-Inner Classes