Add: HashSet
Java HashSet
A HashSet is a collection of items where every item is unique, and it is found in the java.util package.

The HashSet class has many useful methods. For example, to add items to it, use the add() method.

Check If an Item Exists
To check whether an item exists in a HashSet, use the contains() method, even though some object is added twice it only appears once in the set because every item in a set has to be unique.

Check If an Item Exists
To check whether an item exists in a HashSet, use the contains() method.

Remove an Item
To remove an item, use the remove() method, To remove all items, use the clear() method.

To find out how many items there are, use the size method.

Loop Through a HashSet
Loop through the items of an HashSet with a for-each loop.

Other Types
Items in an HashSet are actually objects. In the examples above, we created items (objects) of type "String". Remember that a String in Java is an object (not a primitive type). To use other types, such as int, you must specify an equivalent wrapper class: Integer. For other primitive types, use: Boolean for boolean, Character for char, Double for double, etc.

Resume:
-HashSet